+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * Added Python 3 as a dependency for Homebrew installations
   to ensure some of the newer HTTP features work out of the box
   for macOS users (starting with HTTPie 0.9.4.).
+* Added the ability to unset a request header with ``Header:``, and send an
+  empty value with ``Header;``.
 
 
 `0.9.4`_ (2016-07-01)